Add files via upload
Not so impressive. Some potential signal at 90, but nothing really interesting by 20.  The lesson is that I need more data than one equity. The hypothesis (that RSI at the tail ends anticipates a price reversion) may not be wrong, but we cannot infer anything from this minute sample. The next step is to set up the numbers crunch algorithmically so a lot more can be processed.